feat: リファクタリングのためのディレクトリ構成変更
