+++ lib/migration.ts
@@ -1,3 +1,4 @@
+
 /**
  * データ移行ユーティリティ
  * Cookieからデータベースへのデータ移行
